+++ MaxMind.MinFraud.UnitTest/Request/EmailTest.cs
@@ -173,6 +173,14 @@ namespace MaxMind.MinFraud.UnitTest.Request
             e = new Email(address: "foo@example.comcom", hashAddress: true);
             Assert.Equal("b48def645758b95537d4424c84d1a9ff", e.AddressMD5);
             Assert.Equal("example.comcom", e.Domain);
+
+            e = new Email(address: "foo@example.com.", hashAddress: true);
+            Assert.Equal("b48def645758b95537d4424c84d1a9ff", e.AddressMD5);
+            Assert.Equal("example.com.", e.Domain);
+
+            e = new Email(address: "foo@example.com...", hashAddress: true);
+            Assert.Equal("b48def645758b95537d4424c84d1a9ff", e.AddressMD5);
+            Assert.Equal("example.com...", e.Domain);
         }
 
         [Fact]